requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==10.0.0
pymongo==4.5.0
python-dotenv==1.0.0